+++ sendcards/sendcards.py
@@ -56,22 +56,22 @@ class SendCards(commands.Cog):
         """Send a card to someone!"""
 
     @send.command()
-    async def christmas(self, ctx: commands.Context, user_id: int, *, message: str):
+    async def christmas(self, ctx: commands.Context, user_id: int, *, message: str, image: discord.Attachment = None):
         """Send a christmas card to someone."""
         await self.card_send(ctx, "christmas", user_id, message)
 
     @send.command()
-    async def birthday(self, ctx: commands.Context, user_id: int, *, message: str):
+    async def birthday(self, ctx: commands.Context, user_id: int, *, message: str, image: discord.Attachment = None):
         """Send a birthday card to someone."""
         await self.card_send(ctx, "birthday", user_id, message)
 
     @send.command(aliases=["gws"])
-    async def getwellsoon(self, ctx: commands.Context, user_id: int, *, message: str):
+    async def getwellsoon(self, ctx: commands.Context, user_id: int, *, message: str, image: discord.Attachment = None):
         """Send a get well soon card to someone."""
         await self.card_send(ctx, "get well soon", user_id, message)
 
     @send.command()
-    async def valentine(self, ctx: commands.Context, user_id: int, *, message: str):
+    async def valentine(self, ctx: commands.Context, user_id: int, *, message: str, image: discord.Attachment = None):
         """Send a valentines card to someone."""
         await self.card_send(ctx, "valentines", user_id, message)
 
@@ -101,9 +101,12 @@ class SendCards(commands.Cog):
         embed.set_footer(
             text=f"Send {type} cards by using: {ctx.clean_prefix}send {type.replace(' ', '')}!"
         )
+        if ctx.message.attachments:
+            image = await ctx.message.attachments[0].to_file()
+            embed.set_image(url="attachment://" + str(image.filename))
         try:
             if await ctx.embed_requested():
-                await name.send(embed=embed)
+                await name.send(embed=embed, file=image if ctx.message.attachments else None)
             else:
                 await ctx.send(f"{bold(title)}\n\n{description}")
             return await ctx.send(