+++ sendcards/CHANGELOG.rst
@@ -5,3 +5,9 @@ SendCards: Changelog
 ====================
 
 This changelog includes all changes to the cog since 28/02/2021.
+
+----------
+02/03/2021
+----------
+
+* Allow cards to have attachments inside of them.